+++ main.py
@@ -1,28 +1,9 @@
-from enum import Enum
-from typing import List
-
+from typing import List, Dict, Any
 from aiocache import cached, RedisCache
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from pydantic import BaseModel, PositiveInt, PositiveFloat
 
-
-class Status(Enum):
-    completed = "completed"
-    pending = "pending"
-    canceled = "canceled"
-    all = "all"
-
-
-class Order(BaseModel):
-    id: PositiveInt
-    item: str
-    quantity: PositiveInt
-    price: PositiveFloat
-    status: Status = Status.completed
-
-
-LIST_ORDER_STATUS = dict[str, list[Order] | Status]
+from model.order import Order, Status, LIST_ORDERS, OrdersResponse
 
 app = FastAPI()
 
@@ -48,8 +29,8 @@ async def say_hello(name: str):
 
 
 @cached(cache, key_builder="solution")
-@app.post("/solution", tags=["Solution"])
-async def process_orders(orders: List[Order], criterion: Status) -> LIST_ORDER_STATUS:
+@app.post("/solution", tags=["Solution"], response_model=OrdersResponse)
+async def process_orders(orders: List[Order], criterion: Status) -> LIST_ORDERS:
     filtered_orders = []
     for order in orders:
         if (order.status == criterion) or (criterion == Status.all):
