+++ requeriments.txt
@@ -9,6 +9,7 @@ idna==3.4
 pydantic==1.10.7
 python-dotenv==1.0.0
 PyYAML==6.0
+redis==4.5.4
 sniffio==1.3.0
 starlette==0.26.1
 typing_extensions==4.5.0
